DEV-113428
No offset checks anymore in positioning. It fills the defaultBox and centers the configuration if respectTransform is false.
It respects offset(as a ratio) and scale(as a ratio). An offset leading to an hard boundary comes from a previous hard boundary error in the respectTransform=true case.

productViewer.localToGlobal not properly bounded. When viewBox changed the localToGlobal factor changes but values in the componenets would not be updated. For example the configuration handle radius.